consolidate alerts
