+++ src/applications/debt-letters/components/DebtLettersDownload.jsx
@@ -6,45 +6,10 @@ import Breadcrumbs from '@department-of-veterans-affairs/component-library/Bread
 import { setPageFocus } from '../utils/page';
 import DebtLettersTable from './DebtLettersTable';
 import scrollToTop from 'platform/utilities/ui/scrollToTop';
+import { DownloadLettersAlert } from './Alerts';
 import Telephone, {
-  CONTACTS,
   PATTERNS,
 } from '@department-of-veterans-affairs/component-library/Telephone';
-
-const DownloadLettersAlert = () => (
-  <va-alert
-    class="vads-u-margin-top--4 vads-u-margin-bottom--4"
-    status="warning"
-  >
-    <h3 slot="headline">
-      Downloadable letters have incorrect repayment plan terms
-    </h3>
-    <p className="vads-u-font-size--base vads-u-font-family--sans">
-      We’re sorry. The length of time listed for repayment plans in these
-      letters is too short. Use the letters you get in the mail to find the
-      correct repayment plan terms. If you have any questions, call us at
-      <Telephone contact={'800-827-0648'} className="vads-u-margin-x--0p5" />
-      (or
-      <Telephone
-        contact={'1-612-713-6415'}
-        pattern={PATTERNS.OUTSIDE_US}
-        className="vads-u-margin-x--0p5"
-      />
-      from overseas). We’re here Monday through Friday, 7:30 a.m. to 7:00 p.m.
-      ET. If you have hearing loss, call TTY:
-      <Telephone
-        contact={CONTACTS[711]}
-        pattern={PATTERNS['3_DIGIT']}
-        className="vads-u-margin-left--0p5"
-      />
-      .
-    </p>
-    <p className="vads-u-font-size--base vads-u-font-family--sans">
-      We’re working to fix this problem as fast as we can. Check back soon for
-      updates.
-    </p>
-  </va-alert>
-);
 
 const DebtLettersDownload = ({
   debtLinks,
@@ -52,6 +17,8 @@ const DebtLettersDownload = ({
   isVBMSError,
   hasDependentDebts,
 }) => {
+  const showError = isError || isVBMSError;
+
   useEffect(() => {
     scrollToTop();
     setPageFocus('h1');
@@ -67,7 +34,6 @@ const DebtLettersDownload = ({
           Download debt letters
         </a>
       </Breadcrumbs>
-
       <div className="large-screen:vads-l-col--8">
         <h1
           id="downloadDebtLetters"
@@ -80,15 +46,12 @@ const DebtLettersDownload = ({
           Download your debt letters, learn your payment options, or find out
           how to get help with your VA debts.
         </p>
-
         <DownloadLettersAlert />
-
         <h2>Your debt letters</h2>
         <DebtLettersTable
           debtLinks={debtLinks}
-          isError={isError}
-          isVBMSError={isVBMSError}
           hasDependentDebts={hasDependentDebts}
+          isError={showError}
         />
         <div className="vads-u-margin-bottom--6 vads-u-margin-top--5">
           <h2 className="vads-u-margin-y--0">
